error catching added to user routes

+++ RogueBookingsServer/src/main/java/RogueBookings/services/UserService.java
@@ -1,6 +1,7 @@
 package RogueBookings.services;
 
 import RogueBookings.converters.DTOConverter;
+import RogueBookings.exception.OopsieRequestException;
 import RogueBookings.models.User;
 import RogueBookings.models.userLogs.UserLogs;
 import RogueBookings.repositories.UserLogsRepository;
@@ -9,10 +10,13 @@ import RogueBookings.dataTransferObjects.UserDTO;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.lang.reflect.Type;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -22,6 +26,9 @@ public class UserService {
     DTOConverter<UserDTO, User> dtoConverter;
     Type userDTOType = new TypeToken<UserDTO>() {}.getType();
     Type userType = new TypeToken<User>() {}.getType();
+
+    private User user;
+    private UserLogs userLogs;
     UserLogsRepository userLogsRepository;
 
 
@@ -31,28 +38,48 @@ public class UserService {
         this.dtoConverter = dtoConverter;
         this.userLogsRepository = userLogsRepository;
         this.modelMapper = new ModelMapper();
+
         modelMapper.getConfiguration().setSkipNullEnabled(true);
     }
 
     public List<UserDTO> getAllUsers() {
         return dtoConverter.entityToDTO(userRepository.findAll(), userDTOType);
     }
-    public void addNewUser(User user) {
-        UserLogs userLogs = new UserLogs();
+    public UserDTO addNewUser(UserDTO userDTO) {
+        user = dtoConverter.DTOtoEntity(userDTO, userType);
+        userLogs = new UserLogs();
         userLogs.setUser(user);
         user.setUserLogs(userLogs);
-        userRepository.save(user);
+        try {
+            return dtoConverter.entityToDTO(userRepository.save(user), userDTOType);
+        }catch (Exception e){
+            throw new OopsieRequestException("save did not work try again lol");
+        }
     }
 
-    public void deleteUserById(Long userId) {
+    public ResponseEntity<String> deleteUserById(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new OopsieRequestException("Oopsie This user doesnt exist sadge");
+        }
         userRepository.deleteById(userId);
+        return new ResponseEntity<>("Successfully deleted user", HttpStatus.OK);
     }
 
-    public User editUser(UserDTO userDTO, Long userId) {
-        User patchObj = dtoConverter.DTOtoEntity(userDTO, userType);
-        User user = userRepository.findById(userId).get();
-        modelMapper.map(patchObj, user);
-        userRepository.save(user);
-        return user;
+    public UserDTO editUser(UserDTO userDTO, Long userId) {
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isPresent()){
+            User userToUpdate = user.get();
+            User updatedUser;
+                try {
+                    User patchObj = dtoConverter.DTOtoEntity(userDTO, userType);
+                    modelMapper.map(patchObj, userToUpdate);
+                    updatedUser = userRepository.save(userToUpdate);
+                }catch (Exception e){
+                    throw new OopsieRequestException("Edits did not work sry");
+                }
+                return dtoConverter.entityToDTO(updatedUser, userDTOType);
+        }
+        throw new OopsieRequestException("Oopsie This user doesnt exist sadge");
     }
 }
+
